Update README.md
# Food Site
This is a functional website with some different food menu options that will allow you to view an image display, price, the name of the food, a link for details of the item, an option to add it to the cart and also to checkout. You can also add or delete items to the site, register user profiles, and log in and out of those profiles. All of this can of course be managed as well from the django admin.